requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.25.1
beautifulsoup4==4.9.3
nltk==3.6.5
googlesearch-python==1.1.0
gunicorn==20.1.0
